requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
httpx==0.26.0
sqlalchemy>=2.0.36
apscheduler==3.10.4
twilio==9.0.0